// src/generated/prisma/enums.ts
export const UserStatus = {
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED',
  SELLING_RESTRICTED: 'SELLING_RESTRICTED',
  WITHDRAWAL_HOLD: 'WITHDRAWAL_HOLD',
  BANNED: 'BANNED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]